Initial checkin of the PKCS #11 based ACVP system.    Only the setup has been tested.    AEAD has not been implemented    DRBG has not been implemented (will be plaform specific)    AES_KEY_WRAP has not been implemented
This app has been built to work in NSS softoken, but it has been put together
so that it should work with most tokens. Because softoken is used by curl
on RHEL7, the module under test is loaded in a separate process and the
PKCS #11 commands are marcshalled between the two processes. pkcs11_server.c
handles this.